% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelize_kobo.R
\name{apply_labels}
\alias{apply_labels}
\title{Aplly SPSS .sps label to variable modality name coded}
\usage{
apply_labels(data, spsFilePath)
}
\arguments{
\item{data}{Unlabel value label data read in R}

\item{spsFilePath}{Path to the .sps file}
}
\value{
A dataframe `data` (data frame with variables labelled)
}
\description{
Aplly SPSS .sps label to variable modality name coded
}
