% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelize_kobo.R
\name{labeliser_variables}
\alias{labeliser_variables}
\title{Labelize variable with SPSS .sps label}
\usage{
labeliser_variables(data, spsFilePath)
}
\arguments{
\item{data}{Unlabel data read in R}

\item{spsFilePath}{Path to the .sps file}
}
\value{
A dataframe `data` (data frame with variables labelled)
}
\description{
Labelize variable with SPSS .sps label
}
