% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelize_kobo.R
\name{labelize_my_data}
\alias{labelize_my_data}
\title{labelize kobotoolbox,ODK, ONA data to data labelled including answers code and variable label}
\usage{
labelize_my_data(unlabel.data, labelfile.path)
}
\arguments{
\item{unlabel.data}{: unlabel data (it may be a dataframe)}

\item{labelfile.path}{: label file path (.sps)}
}
\value{
A dataframe `data` (data frame with variables and modalities labelled)
}
\description{
labelize kobotoolbox,ODK, ONA data to data labelled including answers code and variable label
}
