% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelize_kobo.R
\name{parser_labels}
\alias{parser_labels}
\title{Parse label text and return a list of value labels by variable}
\usage{
parser_labels(label_text)
}
\arguments{
\item{label_text}{A character vector of lines containing labels}
}
\value{
A named list: each element is a list of value-label pairs for a variable
}
\description{
Parse label text and return a list of value labels by variable
}
