% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelize_kobo.R
\name{process_sps_file}
\alias{process_sps_file}
\title{Process SPSS .sps label file}
\usage{
process_sps_file(file_path)
}
\arguments{
\item{file_path}{Path to the .sps file}
}
\value{
A list with `valueLabel` (character vector) and `labels_df` (data frame of variable labels)
}
\description{
Process SPSS .sps label file
}
