% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelize_kobo.R
\name{transform_to_label_text}
\alias{transform_to_label_text}
\title{Transform a vector of label lines into a single text block}
\usage{
transform_to_label_text(input_vector)
}
\arguments{
\item{input_vector}{A character vector containing label lines}
}
\value{
A single string with newline-separated labels
}
\description{
Transform a vector of label lines into a single text block
}
